+++ src/App.js
@@ -2,7 +2,7 @@
 import {
   BrowserRouter, Routes, Route, Outlet,
 } from 'react-router-dom';
-import './App.css';
+import './styles/App.css';
 import Home from './Routes/HomeScreen';
 import Quote from './Routes/QuoteScreen';
 import Calculator from './Routes/CalculatorScreen';
