+++ src/styles/App.scss
